feat: Improve example variables (#131)
* feat: add no-bootstrap flag and use prameter expansion

* add a zilla version param

* update naming to init-kakfa

* use correct namespace

* update to use kafka-server instead of seprate host and port

* use options.server in kafka clients

+++ examples/mqtt.kafka.asyncapi.proxy/docker/compose/setup.sh
@@ -1,26 +1,28 @@
 #!/bin/bash
 set -e
 
-if [[ -z "$KAFKA_HOST" && -z "$KAFKA_PORT" ]]; then
-  export KAFKA_HOST=host.docker.internal
-  export KAFKA_PORT=9092
-  echo "==== This example requires env vars KAFKA_HOST and KAFKA_PORT for a running kafka instance. Setting to the default ($KAFKA_HOST:$KAFKA_PORT) ===="
-fi
-
-NAMESPACE=zilla-mqtt-kafka-asyncapi-proxy
+NAMESPACE="${NAMESPACE:-zilla-mqtt-kafka-asyncapi-proxy}"
+export ZILLA_VERSION="${ZILLA_VERSION:-latest}"
+export KAFKA_BROKER="${KAFKA_BROKER:-kafka}"
+export KAFKA_BOOTSTRAP_SERVER="${KAFKA_BOOTSTRAP_SERVER:-host.docker.internal:9092}"
+export KAFKA_PORT="${KAFKA_PORT:-9092}"
+INIT_KAFKA="${INIT_KAFKA:-true}"
 
 # Start or restart Zilla
 if [[ -z $(docker-compose -p $NAMESPACE ps -q zilla) ]]; then
+  echo "==== Running the $NAMESPACE example with $KAFKA_BROKER($KAFKA_BOOTSTRAP_SERVER) ===="
   docker-compose -p $NAMESPACE up -d
 
   # Create the mqtt topics in Kafka
-  docker run --rm bitnami/kafka:3.2 bash -c "
-  echo 'Creating topics for $KAFKA_HOST:$KAFKA_PORT'
-  /opt/bitnami/kafka/bin/kafka-topics.sh --bootstrap-server $KAFKA_HOST:$KAFKA_PORT --create --if-not-exists --topic mqtt-messages
-  /opt/bitnami/kafka/bin/kafka-topics.sh --bootstrap-server $KAFKA_HOST:$KAFKA_PORT --create --if-not-exists --topic streetlights
-  /opt/bitnami/kafka/bin/kafka-topics.sh --bootstrap-server $KAFKA_HOST:$KAFKA_PORT --create --if-not-exists --topic mqtt-retained --config cleanup.policy=compact
-  /opt/bitnami/kafka/bin/kafka-topics.sh --bootstrap-server $KAFKA_HOST:$KAFKA_PORT --create --if-not-exists --topic mqtt-sessions --config cleanup.policy=compact
-  "
+  if [[ $INIT_KAFKA == true ]]; then
+    docker run --rm bitnami/kafka:3.2 bash -c "
+    echo 'Creating topics for $KAFKA_BOOTSTRAP_SERVER'
+    /opt/bitnami/kafka/bin/kafka-topics.sh --bootstrap-server $KAFKA_BOOTSTRAP_SERVER --create --if-not-exists --topic mqtt-messages
+    /opt/bitnami/kafka/bin/kafka-topics.sh --bootstrap-server $KAFKA_BOOTSTRAP_SERVER --create --if-not-exists --topic streetlights
+    /opt/bitnami/kafka/bin/kafka-topics.sh --bootstrap-server $KAFKA_BOOTSTRAP_SERVER --create --if-not-exists --topic mqtt-retained --config cleanup.policy=compact
+    /opt/bitnami/kafka/bin/kafka-topics.sh --bootstrap-server $KAFKA_BOOTSTRAP_SERVER --create --if-not-exists --topic mqtt-sessions --config cleanup.policy=compact
+    "
+  fi
 
 else
   docker-compose -p $NAMESPACE restart --no-deps zilla
